Fill in `Overview.md`
This fills in `Overview.md` with content from the design issue, CG presentation,
and subsequent discussions.

It tentatively specifies that custom page sizes must be a power of two between 1
and 64KiB. It is worth noting, however, that there is still somewhat active
discussion surrounding what the exact values allowed should be. Powers of two
seems like the option with the most support currently. It is also a superset of
the other options so, should we decide to restrict custom page sizes to only the
values 1 and 64KiB in the future, for example, it should be an easy migration.
See https://github.com/WebAssembly/custom-page-sizes/issues/2 for more details.

There is a section that summarizes the discussion in
https://github.com/WebAssembly/custom-page-sizes/issues/3. I think we've
resolved the toolchain-related open questions we've thought of thus far, so I
will go ahead and say that this commit

Fixes https://github.com/WebAssembly/custom-page-sizes/issues/3

Finally, there is a new section on binary encoding. I have not attempted to
sketch this out yet. I will open a dedicated issue for discussing this subject.