Fill out the validation and execution sections of the overview
This commit fills out the `TODO` placeholders for the details of validation and
execution in the overview. I think everything here should be pretty much what
folks expect. The only surprise, perhaps, is the special casing to disallow growing a
memory that has a page size of `1` to length `2**32 - 1` to avoid ambiguities on
the result of `memory.grow` instructions.

I believe I've enumerated everywhere in the spec that page sizes are referenced;
actually making the spec changes should be pretty easy after this change. But
first, I wanted to get the overview document into a place where it is pretty
much complete.