Adjust missing import to invalid import in memory_max.wast
When the import is missing, the "read the imports" step in
the JS API will cause a TypeError to be thrown instead of
a link error. Changing the condition to an invalid import
with the wrong type causes a LinkError to be thrown instead,
which allows the use of assert_unlinkable.